refactor(rubocop): extract synchronizer rollback concern
Issue: No ticket/issue
